+++ services/gateway/apis/socketio-jsonrpc-version1/methods/blocks.js
@@ -14,7 +14,7 @@
  *
  */
 const blocksSource = require('../../../sources/version1/blocks');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
