+++ services/gateway/apis/socketio-jsonrpc-version1/methods/delegatesNextForgers.js
@@ -14,7 +14,7 @@
  *
  */
 const delegatesSource = require('../../../sources/version1/delegates');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
