+++ services/gateway/apis/socketio-jsonrpc-version1/methods/networkSearch.js
@@ -15,7 +15,7 @@
  */
 
 const networkSearchSource = require('../../../sources/version1/networkSearch');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
