+++ services/gateway/apis/socketio-jsonrpc-version1/methods/networkStatistics.js
@@ -15,7 +15,7 @@
  */
 
 const networkStatisticsSource = require('../../../sources/version1/networkStatistics');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
