+++ services/gateway/apis/socketio-jsonrpc-version1/methods/peers.js
@@ -14,7 +14,7 @@
  *
  */
 const peersSource = require('../../../sources/version1/peers');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 
 module.exports = {
