+++ services/gateway/apis/socketio-jsonrpc-version1/methods/transactions.js
@@ -14,7 +14,7 @@
  *
  */
 const transactionsSource = require('../../../sources/version1/transactions');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
