+++ services/gateway/apis/socketio-jsonrpc-version1/methods/transactionsStatisticsDay.js
@@ -14,7 +14,7 @@
  *
  */
 const transactionsStatisticsDaySource = require('../../../sources/version1/transactionsStatisticsDay');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
