+++ services/gateway/apis/socketio-jsonrpc-version1/methods/voters.js
@@ -14,7 +14,7 @@
  *
  */
 const votersSource = require('../../../sources/version1/voters');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
