+++ services/gateway/apis/socketio-jsonrpc-version1/methods/votes.js
@@ -14,7 +14,7 @@
  *
  */
 const votesSource = require('../../../sources/version1/votes');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
